Use JSONPath to map OpenIDConnect claims to roles and org name
Regardless of the authorization source (LDAP, OAuth2, OIDC),
user credentials are mapped to GeorchestraUser, from which
the sec-* headers will later on be constructed and conveyed
to the back-end georchestra services.

This patch allows to extract the roles and organization
short name from OpenIDConnect standard or non-standard
claims, using a JSONPath expression.